requirements.txt
bcrypt==4.0.1
blinker==1.6.3
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
greenlet==3.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
nltk==3.8.1
PyJWT==2.8.0
python-dotenv==1.0.0
regex==2023.10.3
SQLAlchemy==2.0.22
textblob==0.17.1
tqdm==4.66.1
typing_extensions==4.8.0
uuid==1.30
Werkzeug==3.0.1
WTForms==3.1.0
